backend/requirements.txt
# AllôDoto Backend — Dépendances Python
# Généré pour Python 3.14

# Framework
django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4

# Authentification JWT
djangorestframework-simplejwt>=5.3

# Base de données
psycopg2-binary>=2.9

# Variables d'environnement
python-decouple>=3.8

# Documentation API (Swagger/OpenAPI)
drf-spectacular>=0.27

# Traitement images/fichiers
Pillow>=11.0

# Cache Redis
django-redis>=5.4

# Celery (tâches asynchrones : SMS, notifications)
celery>=5.4
redis>=5.2
